:truck: Move everything up a level
Now that I've dropped the need for locally-hosting non-app code in the
package (originally I had the history code, which is now in bagofstuff, and
the gemtext parser, which I've moved to its own library) there's no need to
have the app itself be in its own sub-directory. This commit promotes
everything in the app up a level.
