+++ src/rogallo/__main__.py
@@ -2,7 +2,7 @@
 
 ##############################################################################
 # Local imports.
-from .app import Rogallo
+from .rogallo import Rogallo
 
 
 ##############################################################################
