--- src/rogallo/app/__init__.py
@@ -1,12 +1,0 @@
-"""The application."""
-
-##############################################################################
-# Local imports.
-from .rogallo import Rogallo
-
-##############################################################################
-# Exports.
-__all__ = ["Rogallo"]
-
-
-### __init__.py ends here
