+++ src/rogallo/data/locations.py
