+++ src/rogallo/data/navigation_history.py
