+++ src/rogallo/messages/opening.py
