+++ src/rogallo/preflight.py
