+++ src/rogallo/types.py
