+++ src/rogallo/widgets/command_line/base_command.py
