+++ src/rogallo/widgets/command_line/general.py
