+++ src/rogallo/widgets/command_line/open_gemini_uri.py
