+++ src/rogallo/widgets/command_line/open_other_uri.py
