+++ src/rogallo/widgets/command_line/widget.py
