+++ src/rogallo/widgets/viewer/__init__.py
