+++ src/rogallo/widgets/viewer/document_view.py
