+++ src/rogallo/widgets/viewer/status.py
