+++ src/rogallo/widgets/viewer/title.py
