+++ src/rogallo/widgets/viewer/widget.py
